feat: Update package references and add JSDoc comments
This commit updates the package references in the README.md and jsr.json
files to reflect the new package name '@ryoppippi/unplugin-typia'. It also
bumps the version number to 0.1.1 in the package.json and jsr.json files.

In addition, JSDoc comments have been added to the index.ts file to provide
more detailed information about the exported modules and functions.