fix( react vite examples ): minor fix
This commit includes the addition of UUID types to the project's
dependencies. It also improves the display of validation results in the
InputForm component. Now, instead of relying on the truthiness of the
result's success property, it explicitly checks for true. In case of
validation failure, it ensures that the error message is displayed
properly.