feat: add typia transformer options to unplugin-typia
This commit introduces the ability to pass transformer options to
typia in the unplugin-typia plugin. The options are added to the
Options interface and are resolved in the resolveOptions function.
The process.cwd() default value for the cwd option is now set in
resolveOptions. The typia transformer is updated to use the new
options. The Options type is also exported from the main index file.