feat(renovate): adjust stabilityDays and add schedule
The stabilityDays for the overall project has been reduced from 22 to 2.
A timezone of Europe/London has been set and a schedule of before 2pm on
Monday has been added. For the packages typia and ts-patch, the schedule
has been set to before 4am. No change in stabilityDays for these packages.