feat: add tsconfig path option to plugin options
This commit introduces a new option 'tsconfig' in the plugin options.
This option allows users to specify the path to the tsconfig file.
If not specified, the plugin will try to find it automatically.
The 'tsconfig' option is also added to the 'ResolvedOptions' type.
The 'getTsCompilerOption' function now accepts an optional 'tsconfigId'
parameter, which is used when reading the tsconfig file.