feat: Refactor bunTypiaPlugin to use unplugin
This commit refactors the `bunTypiaPlugin` function to use the `unplugin`
module instead of the previous `transformTypia` function. The `transform`
function from `unplugin.raw` is now used to transform the source code.

The return type of the `transform` function is also handled differently.
If the result is `null` or a string, the original source is returned.
Otherwise, the transformed code from the result is returned.

This change improves the flexibility and maintainability of the code by
leveraging the `unplugin` module.