feat: update package version and set private
This commit does two main things. First, it updates the version of
the package from 0.3.1 to 0.3.2 in both jsr.json and package.json
files. This is a minor version update, indicating some improvements
or fixes have been made.

Second, it sets the 'private' field in the main package.json to true.
This is usually done to prevent accidental publication of a package
that is meant to be private.