feat: add format script to package.json
A new script 'format' has been added to the scripts section of
package.json in both the root directory and the unplugin-typia
package. This script will run the 'eslint --fix .' command to
automatically format the code according to the project's ESLint
rules. The turbo.json file has also been updated to include a
dependency for the 'format' script.