feat: Simplify build process and package structure
This commit simplifies the build process by using 'tsup' instead of
separate CommonJS and ESM builds. The 'dist-cjs' and 'dist-esm' folders
are replaced with a single 'dist' folder. The package name is also
simplified, and the 'publish' script is updated to reflect these changes.
The 'tsconfig.cjs.json' file is removed as it's no longer needed, and a
new 'tsup.config.ts' file is added for the 'tsup' configuration.