feat: Add GitHub workflow for publishing package
This commit introduces a new GitHub workflow named 'jsr'. This workflow
is triggered on push events for tags and on workflow dispatch. It sets
up the environment with the latest version of 'bun', installs dependencies,
fetches the version from the tag, updates the version in package.json and
jsr.json, and finally publishes the package if the ref is a version tag.