feat: Add GitHub workflow for jsr
This commit introduces a new GitHub workflow named 'jsr'. It triggers
on push events for tags starting with "v". The workflow has read
permissions for contents and write permissions for id-token. It
runs on ubuntu-latest and includes steps for checkout, install
dependencies with frozen lockfile, and publish from the
'packages/unplugin-typia' directory.