feat(README): add limitation note for script usage
This commit adds a note about a limitation when using the script. Users
need to import the `typia` module as a default import in their script.
Type named import is fine. This is due to a bug in `Bun`. A link to the
related issue is also provided for more context.