refactor: Simplify typia path resolution in bun.ts
This commit simplifies the resolution of the typia path in bun.ts by
removing the resolveTypiaPath function and related code. The function
was previously used to resolve the absolute path of the typia module,
but this is no longer necessary.

In addition, a check for CommonJS syntax in the typia library has been
added. This is a temporary solution and should be removed once the
related issue (https://github.com/oven-sh/bun/issues/11783) is resolved.