feat(examples/bun): Add std-env to manage cache in CI environment
This commit introduces the `std-env` package to the project. It is used
to check if the current environment is a Continuous Integration (CI)
environment. The cache option of the UnpluginTypia plugin is now
dynamically set based on the environment. This change helps to optimize
the caching behavior of the plugin in different environments.