feat: update GitHub Actions workflow
This commit updates the GitHub Actions workflow for the project. It
changes the permissions for the contents from read to write. It also
adds a new job named 'release' which is dependent on the 'jsr' job.
This new job sets up the environment, fetches the repository with a
depth of 0, sets up bun with the latest version, and runs the
'bunx changelogithub' command with the GITHUB_TOKEN environment
variable.